pyproject.toml
[build-system]
requires = ["uv_build >= 0.9.10, <0.10.0"]
build-backend = "uv_build"

[project]
name = "seatable_api_async"
version = "0.2.0"
description = "异步版本的 SeaTable API Python 客户端，基于 aiohttp 实现"
authors = [
    { name = "bo-john", email = "boxiong74@gmail.com" },
    { name = "aifunc", email = "tabshiftctrlalt@outlook.com" }
]
readme = "README.md"
license = { text = "Apache License 2.0" }
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.13.2",
    "dotenv>=0.9.9",
    "ply>=3.11",
    "python-dateutil>=2.9.0.post0",
    "python-socketio>=5.11.0",
]

[project.urls]
Repository = "https://github.com/bo-john/seatable_api_async.git"
Issues = "https://github.com/bo-john/seatable_api_async/issues"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
